use flat flatter simpler structure
