--- src/3.linkedlist/src/linkedlist.c
@@ -1,7 +1,0 @@
-#include "linkedlist.h"
-#include <stdlib.h>
-
-linkedlist new(datatype type) {
-    linkedlist temp = {NULL, 0, type};
-    return temp;
-}
