--- src/3.linkedlist/src/node.h
@@ -1,9 +1,0 @@
-#ifndef NODE_H
-#define NODE_H
-
-typedef struct node
-{
-    void *value;
-    struct node *next;
-} node;
-#endif
