+++ src/array/def.txt
