+++ src/array/test-program-take-too-much-memory.c
