+++ src/memory_cell/def.txt
