+++ src/memory_cell/impl
