+++ src/memory_cell/memory-allocation-dynamic.c
