+++ src/memory_cell/memory-allocation-static.c
