+++ src/memory_cell/memory_allocator.c
