+++ src/memory_cell/memory_allocator.h
